已报名活动详情提交

+++ app/src/main/java/com/mango/leo/zsproject/eventexhibition/bean/EventBean.java
@@ -310,7 +310,7 @@ public class EventBean {
             private String name;
             private LocationBean location;
             private Object description;
-            private int price;
+            private Double price;
             private long startTime;
             private long endTime;
             private boolean popular;
@@ -356,11 +356,11 @@ public class EventBean {
                 this.description = description;
             }
 
-            public int getPrice() {
+            public Double getPrice() {
                 return price;
             }
 
-            public void setPrice(int price) {
+            public void setPrice(Double price) {
                 this.price = price;
             }
